python-ai/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.10.2
watchdog==6.0.0
radon==6.0.1
requests==2.32.3
python-multipart==0.0.17
